+++ lib/core/repositories/transfer_repository.dart
@@ -1,5 +1,5 @@
 import 'package:mybudget/models/transfer_model.dart';
-import 'package:objectbox/objectbox.dart';
+import 'package:mybudget/objectbox.g.dart';
 
 class TransferRepository {
   final Box<TransferModel> _box;
@@ -8,6 +8,34 @@ class TransferRepository {
 
   List<TransferModel> getAll() {
     return _box.getAll();
+  }
+
+  TransferModel? get(int id) {
+    return _box.get(id);
+  }
+
+  List<TransferModel> getActive() {
+    final query = _box.query(TransferModel_.endDate.isNull()).build();
+    final results = query.find();
+    query.close();
+    return results;
+  }
+
+  List<TransferModel> getClosed() {
+    final query = _box.query(TransferModel_.endDate.notNull()).build();
+    final results = query.find();
+    query.close();
+    return results;
+  }
+
+  List<TransferModel> getChain(int rootId) {
+    final query = _box
+        .query(TransferModel_.parentId.equals(rootId) |
+            TransferModel_.id.equals(rootId))
+        .build();
+    final results = query.find();
+    query.close();
+    return results;
   }
 
   int add(TransferModel transfer) {